+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ master: 1.9.2
 -------------
 
 * Added ``cppyy.types`` module for exposing cppyy builtin types
+* Improve numpy integration with custom ``__array__`` methods
 
 
 2020-11-23: 1.9.1